+++ client/src/commands.ts
@@ -10,6 +10,8 @@ export { openCompiled } from "./commands/open_compiled";
 export { switchImplIntf } from "./commands/switch_impl_intf";
 export { dumpDebug, dumpDebugRetrigger } from "./commands/dump_debug";
 export { dumpServerState } from "./commands/dump_server_state";
+export { pasteAsRescriptJson } from "./commands/paste_as_rescript_json";
+export { pasteAsRescriptJsx } from "./commands/paste_as_rescript_jsx";
 
 export const codeAnalysisWithReanalyze = (
   targetDir: string | null,